tests/requirements.txt
# Python Testing Framework Dependencies

# Core testing
pytest>=7.4.0
pytest-timeout>=2.1.0
pytest-xdist>=3.3.1  # For parallel test execution
pytest-html>=3.2.0  # For HTML reports

# Kubernetes client
kubernetes>=27.2.0
pyyaml>=6.0.1

# Kafka client
kafka-python>=2.0.2
confluent-kafka>=2.2.0

# HTTP client for Flink REST API
requests>=2.31.0

# Utilities
retry>=0.9.2
tenacity>=8.2.3  # Retry with exponential backoff

# Logging
colorlog>=6.7.0

# Data validation
jsonschema>=4.19.0
